move calculateConfig into lib/utils
