+++ lib/utils.js
@@ -5,6 +5,14 @@
 const fs = require('fs');
 const path = require('path');
 
+const CSP_SELF = "'self'";
+const CSP_NONE = "'none'";
+
+const CSP_HEADER = 'Content-Security-Policy';
+
+const DELIVERY_HEADER = 'header';
+const DELIVERY_META   = 'meta';
+
 const unique = function(array) {
   return array.filter(function(value, index, self) {
     return self.indexOf(value) === index;
@@ -49,7 +57,67 @@ const readConfig = function(project, environment) {
   return fs.existsSync(configPath) ? require(configPath)(environment) : {};
 };
 
+/**
+ * Calculates the configuration based on
+ * - own config (`config/content-security-policy.js`) and
+ * - run-time config (`config/environment.js`) for legacy support.
+ *
+ * @params {string} environment
+ * @params {object} ownConfig
+ * @params {object} runConfig
+ * @params {object} ui
+ * @returns {object}
+ */
+function calculateConfig(environment, ownConfig, runConfig, ui) {
+  let config = {
+    delivery: [DELIVERY_HEADER],
+    enabled: true,
+    policy: {
+      'default-src':  [CSP_NONE],
+      'script-src':   [CSP_SELF],
+      'font-src':     [CSP_SELF],
+      'connect-src':  [CSP_SELF],
+      'img-src':      [CSP_SELF],
+      'style-src':    [CSP_SELF],
+      'media-src':    [CSP_SELF],
+    },
+    reportOnly: true,
+  };
+
+  // testem requires frame-src to run
+  if (environment === 'test') {
+    config.policy['frame-src'] = CSP_SELF;
+  }
+
+  ui.writeWarnLine(
+    'Configuring ember-cli-content-security-policy using `contentSecurityPolicy`, ' +
+    '`contentSecurityPolicyHeader` and `contentSecurityPolicyMeta` keys in `config/environment.js` ' +
+    'is deprecate and will be removed in v2.0.0. ember-cli-content-security-policy is now configured ' +
+    'using `ember-cli-build.js`. Please find detailed information about new configuration options ' +
+    'in addon documentation at https://github.com/rwjblue/ember-cli-content-security-policy/blob/master/DEPRECATIONS.md.',
+    !runConfig.contentSecurityPolicy || !runConfig.contentSecurityPolicyHeader || !runConfig.contentSecurityPolicyMeta
+  );
+
+  // support legacy configuration options
+  if (runConfig.contentSecurityPolicy) {
+    // policy object is merged not replaced
+    Object.assign(config.policy, runConfig.contentSecurityPolicy);
+  }
+  if (runConfig.contentSecurityPolicyMeta) {
+    config.delivery = [DELIVERY_META];
+  }
+  if (runConfig.contentSecurityPolicyHeader) {
+    config.reportOnly = runConfig.contentSecurityPolicyHeader !== CSP_HEADER;
+  }
+
+  // apply configuration
+  Object.assign(config, ownConfig);
+
+  return config;
+}
+
 module.exports = {
   buildPolicyString,
+  calculateConfig,
   readConfig
 };
